💡 refactor(state): setCart 및 주문 완료 로직을 CartSidebar 내부로 이동하여 컴포넌트 자율성 개선
cart 상태 갱신(setCart)과 주문 완료 처리(completeOrder) 로직을 CartSidebar 내부로 이동시켜 props 전달을 줄이고 컴포넌트의 책임을 명확히 분리함.

변경 내용:
- CartSidebar에서 cartAtom setter 직접 사용하여 setCart prop 제거
- completeOrder 로직을 CustomerPage → CartSidebar로 이동
- 주문 완료 시 cart 초기화까지 CartSidebar 내에서 처리
- App.tsx → CustomerPage → CartSidebar로 전달되던 setCart prop 제거

효과:
- App → CustomerPage props 수: 8 → 7
- CustomerPage → CartSidebar props 수: 6 → 5
- CartSidebar의 기능적 독립성 향상
- 기존 주문 완료 기능은 모두 유지되며 테스트 통과 확인
